Updated calculate positions MATLAB and CUDA files
Went ahead and updated the remaining variable names in the MATLAB
function file that takes care of calling the CUDA kernel and passing the
parameters to it. Now all of the parameters passed to this function
match the variable convention necessary to utilize the Param class.

Also updated a variable name in the CUDA function of calculate forces.
NOTE: Most of the parameters passed to the calculateForces kernel are
unused by the function. We should delete those functions along with the
comments that are no longer used.
 Changes to be committed:
	modified:   pcry.cu
	modified:   vdriadCalculatePositions.m