SAVE PARTICLE CONDITIONS (RESTART)
A complete change was made in Vdriad's dust.m to create a function capable of reading and applying the particles' previous positions, masses, charges, and other characteristics, thereby allowing the program to preserve the particles' conditions. Additionally, the Vionwake underwent some modifications to save the particles' conditions, along with the parameters text file (txt) and Vdriad parameters, to incorporate RESTART as a variable in the code. Finally, debugging was carried out in pcry.cu to prevent particles from behaving erratically due to excessive energy and unbalanced forces (this is still a work in progress, as it takes considerable time for particles to settle).